Move bitstream classes in models module as well.
