+++ dspyce/_testing/safTest.py
@@ -3,7 +3,7 @@ import os
 import unittest
 
 from dspyce import saf
-from dspyce import Item, Collection
+from dspyce.models import Item, Collection
 from dspyce.metadata import MetaDataValue
