+++ dspyce/bitstreams/__init__.py
@@ -1,3 +1,1 @@
-from dspyce.bitstreams.Bundle import Bundle
-from dspyce.bitstreams.Bitstream import Bitstream
-from dspyce.bitstreams.IIIFBitstream import IIIFBitstream
+from dspyce.bitstreams.models import Bundle, Bitstream, IIIFBitstream
