+++ dspyce/statistics/retrieve.py
@@ -1,5 +1,5 @@
 from dspyce.rest.RestAPI import RestAPI
-from dspyce.DSpaceObject import DSpaceObject
+from dspyce.models import DSpaceObject
 
 REPORT_TYPES: tuple = ('TotalVisits', 'TotalVisitsPerMonth', 'TotalDownloads', 'TopCountries', 'TopCities')
 """The allowed report_type values."""
